Begin defining CLI by adding arguments
The forkit module provides the functionality that we would like to
achieve. This change begins to define how we use forkit.

To use forkit, we require that you provide the owner's Github login
(samuelrey) and the name of the repository (forkit).

This change does not connect the module to the CLI. This change does not
contain the complete CLI to come.

Test plan:
From the main project directory, run these commands:
`./app.py -h` -- to see the description of the program and arguments
`./app.py` -- to see that the program fails unless both arguments are
provided
`./app.py samuelrey forkit` -- to see that the program succeeds silently