Handle commandline arguments
This diff makes that --clone and --dev are mutually exclusive arguments. This
is necessary because the --clone pipeline will be a subset of the --dev
pipeline.

This diff makes owner_login and repo_name required arguments. I want to
make it such that extension is easy, but it serves these first use cases
I've drawn up well.

There's room for automating these tests.

Test Plan:
 source virtualenv
 setup secrets
 ./app.py samuelrey forkit
        will fork this repository to your github account! :D
 ./app.py samuelrey forkit [--clone|--dev]
        does nothing at the moment!
 ./app.py samuelrey forkit --clone --dev
        print error message
 ./app.py samuelre forkit
        print error message
 ./app.py samuelrey forki
        print error message