Restructure project
Pull he main module and related test file to the highest level
directory. A major factor leading to this change was that we were no
longer using argparse. With click, we were able to keep the logic and
interface together in one place without ending up with a bloated file.

Test Plan: I really need to do this...